Change unlockGsBallEvent() significantly.
Now it not only unlocks the event, but it also makes it repeatable by removing the GS Ball from the players' inventory first
and resetting all GS Ball related event flags.

Not only does this cover up my screw up with PokeMe64. (it prevents players from having to create a new save
to get access to the GS Ball event if they used one of my previous versions of PokeMe64)

As a bonus, it also makes the event completely repeatable, which is nice!