Move the application of the GenXDistributionPokemon properties to separate functions
This allows us to use the genx_prepareDistributionPokemon() functions followed by GenXGameReader::addPokemon()
instead of GenXGameReader::addDistributionPokemon() in order to actually show the resulting GenXTrainerPokemon in
a stats screen in PokeMe64. After all: the genx_prepareDistributionPokemon() function does things like generate IV,
decide the trainer OT name and ID and apply shinyness chances.

So when showing the resulting pokemon in a stats screen, we need the actual resulting pokemon in order to show the right
stats, OT and shinyness