Add code to consider an empty (zero'd out) save as an invalid checksum
The gen 2 saves contain 2 markers per section (main/backup) to serve as canary values to see whether an actual save is stored.
Because after all: checksum 0 is a valid checksum in some cases. So to distinguish between a valid and a corrupted save, these markers are used.

And now we use it too! This will be necessary to detect that there's no save file in case the save was wiped (or a reproduction cart is used)