Fix mystery gift timer
I forgot that Gen2ClockManager is an instance that is being created which needs to be filled first with latchRTC()

And then I need to use that very same instance to retrieve the values. So the call chain I was using earlier wouldn't
work at all, giving a 0 instead of the day counter.

So, this commit changes that.